Update imported-project-utils and use polysquare_import_external_project.
By doing this we are able to drop a substantial amount of boilerplate
and are able to adopt the Google Mock targets in a platform and generator
agnostic manner.

G*_EXTERNAL_SET_LIBRARY has now changed to G*_LIBRARY_LOCATION. It will
be automatically set upon calling polysquare_import_external_project. To
propagate it to other projects, just forward the GTEST and GMOCK cache
namespaces using polysquare_forward_cache_namespaces.